Remove a set of compilation warnings (detected by g++ 9.2.1), most of them being related to sign/unsign comparisons

+++ lib/Importer/ProtobufLoader.cpp
@@ -325,7 +325,7 @@ Expected<TensorQuantizationParams>
 ProtobufLoader::getUpdatedTQP(int32_t uniqueOffsetIdx) {
   RETURN_ERR_IF_NOT(replaceDummyTQPs_, "replaceDummyTQPs_ was not enabled");
   RETURN_ERR_IF_NOT(
-      uniqueOffsetIdx < updatedTQPs_.size(),
+      uniqueOffsetIdx < int32_t(updatedTQPs_.size()),
       strFormat("Unexpected size of updated TQPs %lu vs. dummy offset %d",
                 updatedTQPs_.size(), uniqueOffsetIdx));
   return updatedTQPs_[uniqueOffsetIdx];
@@ -387,7 +387,7 @@ Expected<TypeRef> ProtobufLoader::loadQuantTy(const std::string &name,
   // of the C2 edge that we loaded to use these quant params in the cctx so we
   // can ue it in the future. The index the name is at represents which unique
   // index it is mapped to.
-  RETURN_ERR_IF_NOT(originNameToTQPMap_->size() == currUniqueOffset_,
+  RETURN_ERR_IF_NOT(int32_t(originNameToTQPMap_->size()) == currUniqueOffset_,
                     "Unexpected size encountered for qparam origin tracking");
   const int32_t thisUniqueOffset = currUniqueOffset_++;
   bool inserted =
